slava's suggestions
