+++ RecoLocalMuon/GEMSegment/plugins/GEMSegmentAlgorithm.h
@@ -12,12 +12,11 @@
  *
  */
 
-#include <DataFormats/GEMRecHit/interface/GEMRecHitCollection.h>
-#include <DataFormats/GEMRecHit/interface/GEMSegment.h>
-#include <Geometry/GEMGeometry/interface/GEMSuperChamber.h>
-// #include <Geometry/GEMGeometry/interface/GEMChamber.h>
-#include <Geometry/GEMGeometry/interface/GEMEtaPartition.h>
-#include <FWCore/Framework/interface/Frameworkfwd.h>
+#include "DataFormats/GEMRecHit/interface/GEMRecHitCollection.h"
+#include "DataFormats/GEMRecHit/interface/GEMSegment.h"
+#include "Geometry/GEMGeometry/interface/GEMSuperChamber.h"
+#include "Geometry/GEMGeometry/interface/GEMEtaPartition.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
 #include <map>
 #include <vector>
