+++ RecoLocalMuon/GEMSegment/plugins/SealModule.cc
@@ -9,10 +9,8 @@
 DEFINE_FWK_MODULE(ME0SegmentProducer);
 DEFINE_EDM_PLUGIN(ME0SegmentBuilderPluginFactory, ME0SegAlgoMM, "ME0SegAlgoMM");
 
-#include <RecoLocalMuon/GEMSegment/plugins/GEMSegmentProducer.h>
 #include <RecoLocalMuon/GEMSegment/plugins/GEMSegmentBuilderPluginFactory.h>
 #include <RecoLocalMuon/GEMSegment/plugins/GEMSegAlgo.h>
 
-DEFINE_FWK_MODULE(GEMSegmentProducer);
 DEFINE_EDM_PLUGIN(GEMSegmentBuilderPluginFactory, GEMSegAlgo, "GEMSegAlgo");
 DEFINE_EDM_PLUGIN(ME0SegmentBuilderPluginFactory, ME0SegAlgo, "ME0SegAlgo");
